Started work on GUI.
Created Button class and outlined how GUI will work. When the menu state changes, a new Button List will be created, containing Button objects. When the screen is clicked, the coordinates of the click will be checked against the button bounds. If a click is registered, the menu state will be set to the destination stored by the button. Then, based on the state, the buttons will be re-created and stored in the list. At the moment there are indivisual functions to draw each menu, but they will hopefully be converted into a single general method.