Solved texture issue. New issue arises
Issue with texture was passing numbers larger than 1 (in this case passing 900, the size in pixels of each texture). This lead to stretching of the texture across slim lines on the quad, making the rest of the quad transparent. Now fixed. Now textures are being drawn poorly onto the quad sides of the head segment. Must rearrange calls and coordinates used.